Separated editor functions to fix build issue. Set project to allow resizing window and run in background.
