+++ Assets/testing/scripts/testing_editor_extensions.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: b76455bb613737540ade1340ebe9e3a6
+timeCreated: 1505584197
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
