+++ Assets/u3d_text_3ngine/scripts/u3d_text_3ngine.cs
@@ -8,7 +8,6 @@ using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
 using TMPro;
-using UnityEditor;
 using UnityEngine;
 using Random = System.Random;
 using Stopwatch = System.Diagnostics.Stopwatch;
@@ -105,25 +104,6 @@ public class u3d_text_3ngine : MonoBehaviour
     /// Unity's random number generator is bad
     /// </summary>
     private static Random r = new Random();
-
-    /// <summary>
-    /// Allows right click adding of a u3d_text_3ngine object in the unity editor
-    /// </summary>
-    /// <param name="menuCommand">The event args from unity</param>
-    [MenuItem("GameObject/UI/u3d_text_3ngine")]
-    static void CreateCustomGameObject(MenuCommand menuCommand)
-    {
-        // Create a custom game object
-        GameObject engine = new GameObject("u3d_text_3ngine");
-        engine.AddComponent<u3d_text_3ngine>();
-
-        // Ensure it gets reparented if this was a context click (otherwise does nothing)
-        GameObjectUtility.SetParentAndAlign(engine, menuCommand.context as GameObject);
-
-        // Register the creation in the undo system
-        Undo.RegisterCreatedObjectUndo(engine, "Create " + engine.name);
-        Selection.activeObject = engine;
-    }
 
     #endregion
