+++ Assets/u3d_text_3ngine/scripts/u3d_text_3ngine_editor_extensions.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: b4caad8303a9ac44db667057bc960cc0
+timeCreated: 1505584249
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
